Moved RandomRule to neo4j-common and fixed IndexSearchTest oddity
