+++ community/common/src/test/java/org/neo4j/helpers/Exceptions.java
